implement conference room update and delete

+++ ConferenceRoomBooking.Service/Services/ConferenceRoomService.cs
@@ -10,12 +10,14 @@ public class ConferenceRoomService : IConferenceRoomService
 {
     private readonly IConferenceRoomRepository _conferenceRoomRepository;
     private readonly IRoomServiceRepository _roomServiceRepository;
+    private readonly IAdditionalServiceRepository _additionalServiceRepository;
 
     public ConferenceRoomService(IConferenceRoomRepository conferenceRoomRepository,
-        IRoomServiceRepository roomServiceRepository)
+        IRoomServiceRepository roomServiceRepository, IAdditionalServiceRepository additionalServiceRepository)
     {
         _conferenceRoomRepository = conferenceRoomRepository;
         _roomServiceRepository = roomServiceRepository;
+        _additionalServiceRepository = additionalServiceRepository;
     }
 
     public async Task<ConferenceRoomDto> GetByIdAsync(int id)
@@ -23,7 +25,7 @@ public class ConferenceRoomService : IConferenceRoomService
         var conferenceRoom = await _conferenceRoomRepository.GetByIdAsync(id);
         if (conferenceRoom == null)
         {
-            throw new ConferenceRoomException("Failed_to_Find", "Failed to Find Conference Room");
+            throw new ConferenceRoomException("Failed_to_Add", $"Failed to Find Conference Room with Id = {id}");
         }
 
         return ConferenceRoomMapper.ToConferenceRoomDto(conferenceRoom);
@@ -51,12 +53,70 @@ public class ConferenceRoomService : IConferenceRoomService
     public async Task UpdateAsync(int id, UpdateConferenceRoomDto conferenceRoom)
     {
         var conferenceRoomToUpdate = await _conferenceRoomRepository.GetByIdAsync(id);
+        if (conferenceRoomToUpdate == null)
+        {
+            throw new ConferenceRoomException("Failed_to_Update", $"Failed to Find Conference Room with Id = {id}");
+        }
+    
+        if (conferenceRoom.Name != null)
+        {
+            conferenceRoomToUpdate.Name = conferenceRoom.Name;
+        }
+
+        if (conferenceRoom.BasePricePerHour != null)
+        {
+            conferenceRoomToUpdate.BasePricePerHour = conferenceRoom.BasePricePerHour.Value;
+        }
+
+        if (conferenceRoom.Capacity != null)
+        {
+            conferenceRoomToUpdate.Capacity = conferenceRoom.Capacity.Value;
+        }
+
+        if (conferenceRoom.AdditionalServicesIds != null)
+        {
+            var existingServiceIds = conferenceRoomToUpdate.RoomServices?
+                .Select(rs => rs.AdditionalServiceId)
+                .ToHashSet();
+            var newServiceIds = conferenceRoom.AdditionalServicesIds
+                .Distinct()
+                .Where(serviceId => existingServiceIds != null && !existingServiceIds.Contains(serviceId))
+                .ToList();
+            foreach (var newServiceId in newServiceIds)
+            {
+                var additionalService = await _additionalServiceRepository.GetByIdAsync(newServiceId);
+                if (additionalService == null)
+                {
+                    throw new ConferenceRoomException("Failed_to_Update_Service",
+                        $"Service with id {newServiceId} Not Found");
+                }
+            }
+
+            if (newServiceIds.Count() > 0)
+            {
+                await _roomServiceRepository.AddRangeAsync(RoomServiceMapper.MapToAddRoomServices(id, newServiceIds));
+            }
+        }
+
         await _conferenceRoomRepository.UpdateAsync(conferenceRoomToUpdate);
     }
 
     public async Task DeleteAsync(int id)
     {
         var conferenceRoomToDelete = await _conferenceRoomRepository.GetByIdAsync(id);
+        if (conferenceRoomToDelete == null)
+        {
+            throw new ConferenceRoomException("Failed_to_Delete", $"Failed to Find Conference Room with Id = {id}");
+        }
+
+        if (conferenceRoomToDelete.RoomServices != null)
+        {
+            foreach (var conferenceRoomService in conferenceRoomToDelete.RoomServices)
+            {
+                await _roomServiceRepository.DeleteAsync(conferenceRoomService);
+            }
+        }
+
         await _conferenceRoomRepository.DeleteAsync(conferenceRoomToDelete);
     }
 }
